Add /opt/puppetlabs/bin to PATH on debian
The PATH to /opt/puppetlabs/bin is not set on debian based systems as it
would be set by puppt-agent.sh in /etc/profile.d/.  On Debian based
systems, /etc/profile.d/puppet-agent.sh is not sourced on no-login
shell. No-login shell is used if you run a remote execution task.
As puppet from puppetlabs is installed in /opt/puppetlabs the REX
job fails